Disabled I18N in settings
set 'USE_I18N = False' in taskbuster/settings/base.py since
maintaining translations along with initial development is likely
impractical. However, all of the infrastructure is in place, so if
it is desireable to add I18N in the future, it will be easy to get
started.

Also updated docs/quickstart.rst to reflect this decision.